+++ Cargo.toml
@@ -8,9 +8,10 @@ rust-version = "1.78.0"
 
 [workspace]
 members = [
-  "reverie",
+  "examples/*",
+  "reverie-engine",
+  "reverie-engine-opengl",
   "reverie-util",
-  "examples/*"
 ]
 
 resolver = "2"
@@ -23,7 +24,7 @@ image = { version = "0.24.9", default-features = false, features = ["png"] }
 newtype-ops = "0.1.4"
 once_cell = "1.19.0"
 raw-gl-context = { version = "0.1.2" }
-reverie-engine = { path = "./reverie" }
+reverie-engine-opengl = { path = "./reverie-engine-opengl" }
 reverie-util = { version = "0.0.6", path = "./reverie-util" }
 serde = { version = "1.0.203", features = ["derive"] }
 serde_json = "1.0.117"
