+++ examples/craft/Cargo.toml
@@ -6,7 +6,7 @@ license.workspace = true
 rust-version.workspace = true
 
 [dependencies]
-reverie-engine.workspace = true
+reverie-engine-opengl.workspace = true
 nalgebra = { version = "0.31.4", features = ["serde-serialize"] }
 nalgebra-glm = "0.17.0"
 c_str_macro.workspace = true
