+++ reverie-engine-opengl/resources/color3d.vert
