+++ reverie-engine/Cargo.toml
@@ -1,10 +1,14 @@
 [package]
-name = "example-raw"
+name = "reverie-engine"
 version.workspace = true
+authors.workspace = true
+categories = ["game-engines"]
+description = "A toy game engine"
 edition.workspace = true
 license.workspace = true
+readme = "../README.md"
+repository.workspace = true
 rust-version.workspace = true
 
 [dependencies]
-c_str_macro.workspace = true
 reverie-engine-opengl.workspace = true
